backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "cryptography>=45.0.0",
    "fastapi>=0.136.0",
    "google-cloud-pubsub>=2.31.1",
    "langfuse>=4.5.0",
    "openai-agents>=0.14.4",
    "psycopg2-binary>=2.9.12",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.13.3",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
